requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv
requests
websockets
twilio
